Revert "Batches"
