+++ .gitignore
@@ -1,6 +1,2 @@
 secrets.toml
 *.pyc
-clientsettings.yaml
-batches/
-*.zip
-.DS_Store
