+++ pyproject.toml
@@ -11,8 +11,7 @@ dependencies = [
     "openai (>=1.91.0,<2.0.0)",
     "python-dotenv (>=1.1.1,<2.0.0)",
     "streamlit (>=1.46.0,<2.0.0)",
-    "pyyaml (>=6.0.2,<7.0.0)",
-    "pydantic (>=2.11.7,<3.0.0)"
+    "pyyaml (>=6.0.2,<7.0.0)"
 ]
 
 [tool.poetry]
