+++ src/alttext-generation/config.py
@@ -4,7 +4,7 @@ from typing import Any
 
 def get_client_settings() -> dict[str,str|int]:
     """Load the client settings from the config file."""
-    config_path = Path(__file__).parent.parent.parent/ "clientsettings.yaml"
+    config_path = Path(__file__).parent / "clientsettings.yaml"
 
     with open(config_path, "r") as f:
         config = yaml.safe_load(f)
